pyproject.toml
[project]
name = "facial-emotion-detection"
version = "0.1.0"
description = "Live facial emotion detection website"
authors = [
    {name = "User"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "tensorflow (>=2.20.0,<3.0.0)",
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "numpy (>=2,<2.3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
